Extract lua related logic into a class
